Update .env.template
File system tools can be accessed at '/storage' from the Docker container root. Ensure that you have uploaded a folder named 'storage' to your Docker container in the Docker engine. Any additional files must be uploaded in folders in the '/storage' folder. For example, you might be working on a analyzing a dataset locally in a folder named "data-analysis". For the filesystem tools to effectively interact with these files, you must upload/import that folder to the 'storage' folder in the active Docker container.